+++ src/configuration.ts
@@ -48,6 +48,7 @@ const configuration = {
   SERVER: {
     PORT: Number(process.env.SERVER_PORT) || 8080,
     VERSION: process.env.npm_package_version || packageJson.version,
+    DEMO_MODE: parseBooleanEnvVariable(process.env.SERVER_DEMO_MODE),
     LOG_LEVEL: process.env.SERVER_LOG_LEVEL || "info",
     LOG_FILES_ENABLED: parseBooleanEnvVariable(
       process.env.SERVER_LOG_FILES_ENABLED,