+++ src/decorators/minimum-role.decorator.ts
@@ -1,4 +1,5 @@
 import { SetMetadata } from "@nestjs/common";
 import { Role } from "../modules/users/models/role.enum";
 
-export const MinimumRole = (role: Role) => SetMetadata("minimumRole", role);
+export const MINIMUM_ROLE_KEY = "minimumRole";
+export const MinimumRole = (role: Role) => SetMetadata(MINIMUM_ROLE_KEY, role);